Decompose RulesDisplay into a module with separate SFCs, allowing for prebuild.
