+++ client/src/components/RulesDisplay/index.js
@@ -1,0 +1,1 @@
+export { RulesDisplay } from "./RulesDisplay.vue";
